refactor: removing validations from repository

+++ app/Services/TaskService.php
@@ -5,7 +5,9 @@ namespace App\Services;
 use App\Http\Repositories\TaskRepository;
 use App\Http\Resources\Comment\CommentResource;
 use App\Http\Resources\Task\TaskResource;
+use Illuminate\Contracts\Queue\EntityNotFoundException;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Support\Facades\Log;
 
 class TaskService {
 
@@ -16,10 +18,16 @@ class TaskService {
     }
 
     public function show(int $id): TaskResource {
+        if(!$this->repository->existsByColumn('id', $id)) {
+            throw new EntityNotFoundException('Task', $id);
+        }
         return new TaskResource($this->repository->show($id));
     }
 
     public function showComments(int $id): AnonymousResourceCollection {
+        if(!$this->repository->existsByColumn('id', $id)) {
+            throw new EntityNotFoundException('Task', $id);
+        }
         return CommentResource::collection($this->repository->show($id)->comments);
     }
 
@@ -28,26 +36,42 @@ class TaskService {
     }
 
     public function storeSubtask(int $id, mixed $data): TaskResource {
+        if(!$this->repository->existsByColumn('id', $id)) {
+            throw new EntityNotFoundException('Task', $id);
+        }
+
         $mainTask = $this->repository->show($id);
         $subTask = $this->repository->store($data);
 
+        Log::info($mainTask->id);
+        Log::info($subTask->id);
         $mainTask->subtasks()->attach($subTask->id);
 
         return new TaskResource($subTask);
     }
 
     public function getAllSubtasks(int $id): AnonymousResourceCollection {
+        if(!$this->repository->existsByColumn('id', $id)) {
+            throw new EntityNotFoundException('Task', $id);
+        }
+
         $task = $this->repository->show($id);
 
         return TaskResource::collection($task->subtasks);
     }
 
     public function update(int $id, mixed $data): TaskResource {
+        if(!$this->repository->existsByColumn('id', $id)) {
+            throw new EntityNotFoundException('Task', $id);
+        }
         $task = $this->repository->show($id);
         return new TaskResource($this->repository->update($task, $data));
     }
 
     public function destroy(int $id): void {
+        if(!$this->repository->existsByColumn('id', $id)) {
+            throw new EntityNotFoundException('Task', $id);
+        }
         $this->repository->destroy($id);
     }
 }
